refactor: MetricReader and its type bounded with traits for better clarity
